loading created

+++ src/hooks/useProducts.js
@@ -5,22 +5,25 @@ export function useProducts() {
   const [products, setProducts] = useState([]);
   const [error, setError] = useState(null);
   const [success, setSuccess] = useState(null);
+  const [loading, setLoading] = useState(null);
 
   useEffect(() => {
     loadProducts();
   }, []);
 
-  useEffect(() => 
-    {
-      if (success){
-        const timer = setTimeout(() => 
+
+  useEffect(() => {
+    if (success) {
+      const timer = setTimeout(() =>
         setSuccess(null),
         3000);
 
-        return () => clearTimeout(timer);
+      return () => clearTimeout(timer);
 
-      }
-    }, [success] )
+    }
+  }, [success])
+
+
 
   async function loadProducts() {
     try {
@@ -28,6 +31,8 @@ export function useProducts() {
       setProducts(data);
     } catch (err) {
       setError("Failed to load products.");
+    } finally {
+      setLoading(false);
     }
   }
 
@@ -40,41 +45,53 @@ export function useProducts() {
     } catch (err) {
       setError("Failed to create product.");
       setSuccess(null);
-
+    }
+    finally {
+      setLoading(false);
     }
   }
 
   async function handleDelete(id) {
+
+    const confirmDelete = window.confirm("Are you sure you want to delete?");
+    if (!confirmDelete) return;
+
     try {
       await deleteProduct(id);
       setProducts((prev) => prev.filter((p) => p.id !== id));
-      setSuccess("Product deleted successfully.")
+      setSuccess("Product deleted successfully ✅")
     } catch (err) {
       setError("Failed to delete product.");
       setSuccess(null);
-    
+    }
+    finally {
+      setLoading(false);
     }
   }
 
   async function handleUpdate(id, product) {
-  try {
-    const updated = await updateProduct(id, product);
+    try {
+      const updated = await updateProduct(id, product);
 
-    setProducts((prev) =>
-      prev.map((p) => (p.id === id ? updated : p))
-    );
-    setError(null);
-    setSuccess("Product updated successfully ✅")
-  } catch (err) {
-    setError("Failed to update product.");
-    setSuccess(null)
+      setProducts((prev) =>
+        prev.map((p) => (p.id === id ? updated : p))
+      );
+      setError(null);
+      setSuccess("Product updated successfully ✅")
+    } catch (err) {
+      setError("Failed to update product.");
+      setSuccess(null)
+    } finally {
+      setLoading(false);
+    }
   }
-}
+
 
   return {
     products,
     error,
     success,
+    loading,
     handleCreate,
     handleDelete,
     handleUpdate
